refactor: reorganize trigger executors and enhance plugin node execution handling
