+++ src/main/java/org/phong/zenflow/plugin/subdomain/executors/builtin/triggers/ScheduleTriggerExecutor.java
@@ -1,0 +1,23 @@
+package org.phong.zenflow.plugin.subdomain.executors.builtin.triggers;
+
+import lombok.extern.slf4j.Slf4j;
+import org.phong.zenflow.plugin.subdomain.execution.dto.ExecutionResult;
+import org.phong.zenflow.plugin.subdomain.execution.interfaces.PluginNodeExecutor;
+import org.phong.zenflow.workflow.subdomain.node_definition.definitions.dto.WorkflowConfig;
+import org.springframework.stereotype.Component;
+
+import java.util.Collections;
+
+@Component
+@Slf4j
+public class ScheduleTriggerExecutor implements PluginNodeExecutor {
+    @Override
+    public String key() {
+        return "core:schedule.trigger:1.0.0";
+    }
+
+    @Override
+    public ExecutionResult execute(WorkflowConfig config) {
+        return ExecutionResult.success(Collections.emptyMap(), Collections.emptyList());
+    }
+}
