+++ src/main/java/org/phong/zenflow/workflow/subdomain/node_definition/definitions/BaseWorkflowNode.java
@@ -1,8 +1,6 @@
 package org.phong.zenflow.workflow.subdomain.node_definition.definitions;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonSubTypes;
-import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -10,7 +8,6 @@ import lombok.Setter;
 import org.phong.zenflow.workflow.subdomain.node_definition.definitions.dto.WorkflowConfig;
 
 import org.phong.zenflow.workflow.subdomain.node_definition.definitions.plugin.PluginNodeIdentifier;
-import org.phong.zenflow.workflow.subdomain.node_definition.definitions.trigger.TriggerNodeDefinition;
 import org.phong.zenflow.workflow.subdomain.node_definition.enums.NodeType;
 
 import java.util.List;
@@ -20,11 +17,6 @@ import java.util.Map;
 @Data
 @AllArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
-@JsonTypeInfo(use = JsonTypeInfo.Id.NAME, include = JsonTypeInfo.As.EXISTING_PROPERTY, property = "type", visible = true)
-@JsonSubTypes({
-        @JsonSubTypes.Type(value = BaseWorkflowNode.class, name = "PLUGIN"),
-        @JsonSubTypes.Type(value = TriggerNodeDefinition.class, name = "TRIGGER")
-})
 public class BaseWorkflowNode {
     @NotNull
     private final String key;
